feat: checks for updates automatically and alerts user - moved it to a new file as a utility function
