+++ src/program.js
@@ -1,7 +1,6 @@
 /* @flow */
 import path from 'path';
 import {readFileSync} from 'fs';
-import updateNotifier from 'update-notifier';
 
 import git from 'git-rev-sync';
 import yargs from 'yargs';
@@ -10,6 +9,7 @@ import defaultCommands from './cmd';
 import {UsageError} from './errors';
 import {createLogger, consoleStream as defaultLogStream} from './util/logger';
 import {coerceCLICustomPreference} from './firefox/preferences';
+import {checkForAutomaticUpdates} from './util/updates';
 
 const log = createLogger(__filename);
 const envPrefix = 'WEB_EXT';
@@ -22,10 +22,15 @@ export class Program {
   yargs: any;
   commands: { [key: string]: Function };
   shouldExitProgram: boolean;
+  checkForUpdates: typeof checkForAutomaticUpdates;
 
   constructor(
     argv: ?Array<string>,
-    {absolutePackageDir = process.cwd()}: {absolutePackageDir?: string} = {}
+    {
+      absolutePackageDir = process.cwd(),
+    }: {
+      absolutePackageDir?: string,
+    } = {}
   ) {
     // This allows us to override the process argv which is useful for
     // testing.
@@ -38,6 +43,7 @@ export class Program {
     // config (See web-ext#469 for rationale).
     const yargsInstance = yargs(argv, absolutePackageDir);
 
+    this.checkForUpdates = checkForAutomaticUpdates;
     this.shouldExitProgram = true;
     this.yargs = yargsInstance;
     this.yargs.strict();
@@ -114,7 +120,7 @@ export class Program {
         throw new UsageError(`Unknown command: ${cmd}`);
       }
 
-      checkForAutomaticUpdates ({
+      this.checkForUpdates ({
         name: 'web-ext',
         version: defaultVersionGetter(absolutePackageDir),
         updateCheckInterval: 1000 * 60 * 60 * 24 * 7,
@@ -139,29 +145,6 @@ export class Program {
       }
     }
   }
-}
-
-type checkForAutomaticUpdatesParams = {
-  name: string,
-  version: string,
-  updateCheckInterval: number,
-};
-
-export function checkForAutomaticUpdates({
-  name,
-  version,
-  updateCheckInterval,
-}: checkForAutomaticUpdatesParams
-) {
-  let pkg = {
-    name: name,
-    version: version,
-  };
-
-  updateNotifier({
-    pkg,
-    updateCheckInterval: updateCheckInterval,
-  }).notify();
 }
 
 // A global variable generated by DefinePlugin, generated in webpack.config.js
@@ -195,6 +178,7 @@ export function main(
     runOptions = {},
   }: Object = {}
 ): Promise<any> {
+
   const program = new Program(argv, {absolutePackageDir});
   // yargs uses magic camel case expansion to expose options on the
   // final argv object. For example, the 'artifacts-dir' option is alternatively
